+++ threedigrid_builder/interface/db.py
@@ -355,6 +355,7 @@ class SQLite:
                     models.ConnectionNode.id,
                     models.ConnectionNode.code,
                     models.ConnectionNode.storage_area,
+                    models.ConnectionNode.initial_waterlevel,
                     models.Manhole.id.label("manhole_id"),
                     models.Manhole.calculation_type,
                     models.Manhole.bottom_level,
@@ -371,6 +372,9 @@ class SQLite:
             )
 
         arr["the_geom"] = self.reproject(arr["the_geom"])
+
+        # replace -9999.0 with NaN in initial_waterlevel
+        arr["initial_waterlevel"][arr["initial_waterlevel"] == -9999.0] = np.nan
 
         return ConnectionNodes(**{name: arr[name] for name in arr.dtype.names})
 
@@ -592,7 +596,7 @@ class SQLite:
                 .as_structarray()
             )
 
-        # Pump capicity is entered as L/s but we need m3/s.    
+        # Pump capicity is entered as L/s but we need m3/s.
         arr["capacity"] = arr["capacity"] / 1000
 
         # transform to a Pumps object