+++ threedigrid_builder/tests/test_gridadmin.py
@@ -49,6 +49,7 @@ def h5_out(tmpdir_factory):
             grid_space=20.0,
             dist_calc_points=25.0,
             kmax=4,
+            initial_waterlevel=-0.4,
         ),
         tables_settings=TablesSettings(
             table_step_size=0.05,